requirements.txt
python-dotenv>=1.0.1
openai>=1.40.0
pypdf>=4.2.0
gradio>=3.50.0,<4.0.0
langchain>=0.1.0
langchain-community>=0.0.10
langchain-openai>=0.0.5
chromadb>=0.4.0
google-cloud-storage>=2.0.0
redis>=5.0.0

# Development dependencies
pytest>=8.0.0
pytest-cov>=5.0.0